+++ lib/rspec_starter/task_context.rb
@@ -1,5 +1,5 @@
 module RspecStarter
-  # TaskContext's are created to when parsing the RspecStater.start block. They hold the args and Task class name. They are
+  # TaskContext's are created to when parsing the RspecStater.start block. They hold the ARGS and Task class name. They are
   # asked to create instances of Task subclasses from this information when it is time to execute. The also resolve the options
   # that each Task is allowed to access.
   class TaskContext < StepContext
@@ -25,35 +25,37 @@ module RspecStarter
 
     private
 
+    def registered_task_options
+      environment.options.registered_task_option(@step_class)
+    end
+
     def build_options
       options = super
-      add_defaults_to_options(options)
-      apply_args_to_options(options)
-      apply_command_line_switches_to_options(options)
+      add_task_option_defaults(options)
+      apply_start_block_option_overrides(options)
+      apply_command_line_switch_overrides(options)
       options
     end
 
-    def add_defaults_to_options(options)
-      registered_options = environment.options.registered_task_option(@step_class)
-      registered_options.each { |option| options.add(option.key, option.default) }
+    # Each Task can define options that are unique to the task, along with default values for those options. Add those
+    # options and values now. (The values may be overridden by the user.)
+    def add_task_option_defaults(options)
+      registered_task_options.each { |option| options.add(option.key, option.default) }
     end
 
-    def apply_args_to_options(options)
-      registered_options = environment.options.registered_task_option(@step_class)
-      dsl_option_names = registered_options.select(&:is_dsl_option?).collect { |option| option.name.to_sym }
-      @requested_args.each do |key, value|
-        next unless dsl_option_names.include?(key)
-
-        options.update(key, value, add_missing: false)
-      end
+    # Users can specify options for each task inside their project's "start" block. These options can override the default
+    # value for the global options that all tasks have, or the task specific options that a task defines by calling
+    # register_task_option. If the user tries to define an option that the task doesn't understand, it is ignored.
+    def apply_start_block_option_overrides(options)
+      @requested_args.each { |option_name, option_value| options.update(option_name, option_value, add_missing: false) }
     end
 
-    def apply_command_line_switches_to_options(options)
-      registered_options = environment.options.registered_task_option(@step_class)
+    # Command line switches can also affect options. Adjust an options that are impacted by command line switches that are
+    # present.
+    def apply_command_line_switch_overrides(options)
       present_switches = environment.options.present_switches
-      registered_options.each do |option|
-        # The switch could be nil for the option, which means the step isn't registering a switch for this option. The step
-        # developer just wants to register an option for the dsl (which is applied earlier).
+      registered_task_options.each do |option|
+        # present_switches is a list of all valid switches that have been supplied by the user.
         if option.switch && present_switches.include?(option.switch)
           options.update(option.key, !option.default, add_missing: false)
         end