+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Issues marked as **(Internal)** mark internal development work. Issues are track
 1. [#106](../../issues/106): Remove Ruby 2.5 support.
 1. [#109](../../issues/109): Remove support for legacy API.
 1. [#112](../../issues/112): Change `--no-xvfb` ref to `--skip-display-server` in help output.
+1. [#114](../../issues/114): Fix `quiet` and `stop_on_problem` not getting overridden through the `start` block.
 
 ## 1.8.1 (Apr 14, 2020)